build: extract properties & project extensions & specify debug mode
