+++ app/src/debug/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name" translatable="false">Speak Touch (DEBUG)</string>
+</resources>
